omega-claw-main/requirements.txt
python-telegram-bot>=20.0
python-dotenv>=1.0.0
cryptography>=41.0.0
pexpect>=4.8.0  # Optional: for interactive mode (OMEGA_INVOKER_MODE=interactive)
